Create Breaking_the_records.py
Maria plays college basketball and wants to go pro. Each season she maintains a record of her play. She tabulates the number of times she breaks her season record for most points and least points in a game. Points scored in the first game establish her record for the season, and she begins counting from there.

For example, assume her scores for the season are represented in the array. Scores are in the same order as the games played.

Given Maria's scores for a season, find and print the number of times she breaks her records for most and least points scored during the season.